requirements.txt
django
django-cors-headers
djangorestframework
pymongo==3.12.3